a sense of "mount"ing con"FUSE"ion
Now that we can do directory listings, you'd think that the next thing
we need to implement (glancing at the HelloFS example) is "read". But I
still face huge gaps in my understanding: the read requests are coming
in with an inode handle like 5919553—

[2016-02-27 13:33:23.250552930] read request: Request { ch:
ChannelSender { fd: 3 }, header: fuse_in_header { len: 80, opcode: 15,
unique: 130, nodeid: 5919553, uid: 1000, gid: 1000, pid: 20722, padding:
0 }, data: [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 16, 0, 0,
0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 128, 0, 0, 0, 0, 0, 0] }; ino:
5919553; fh: 0

But ... where is it getting that number from? (It's base-two logarithm
is about 22.49, it's not near two-to-a-power-of-two, so it's not an
overflow of something.)

I really, really don't know what I'm doing here, and that is really,
really terrifying. I guess my vague thought was that we implement
"lookup" for a name on a directory inode to look up a file inode handle
thing, which in our case would proxy to the object on the "real"
filesystem when we wanted to read it. But that description is pretty
handwavy and the fragment of it that is not mere waving of hands could
be wrong! For the protection of everything I hold dear, I must know:
what is the truth? Look! Think!