read path orientation
The ino 5919553 we saw earlier came from the underlying real filesystem:
`lookup` provides the requisite info to the kernel via the reply.entry
call, which is the FileAttr struct I keep inside of my Inode struct,
which the kernel or anyone else has no way of knowing about except
insofar as I write the code to interface them.

This still doesn't work (we're hitting the "couldn't find inode" match
arm), but the world—and the chances of the market-gods of the global
economy continuing to judge me worthy of life—feels less hopeless now.