Have you tried not using inodes? (rewrite/redesign/reimagining)
This hack design is much less general than actually modeling filesystem
structures ourselves, but seems likely to suffice for the immediate goal
of this project (a read-only FUSE filesystem that exposes Markdown files
as HTML), and there may yet be a kind of virtue in this terrible
pragmatism.

This typechecks and compiles, but immediately crashes, because we end up
unwrapping an Option when looking up a file that doesn't exist.